[chore][pkg/stanza] Cleanup remove operator files
