+++ pkg/stanza/operator/transformer/remove/config.go
@@ -4,7 +4,6 @@
 package remove // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/transformer/remove"
 
 import (
-	"context"
 	"fmt"
 
 	"go.uber.org/zap"
@@ -54,34 +53,4 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 		TransformerOperator: transformerOperator,
 		Field:               c.Field,
 	}, nil
-}
-
-// Transformer is an operator that deletes a field
-type Transformer struct {
-	helper.TransformerOperator
-	Field rootableField
-}
-
-// Process will process an entry with a remove transformation.
-func (p *Transformer) Process(ctx context.Context, entry *entry.Entry) error {
-	return p.ProcessWith(ctx, entry, p.Transform)
-}
-
-// Transform will apply the remove operation to an entry
-func (p *Transformer) Transform(entry *entry.Entry) error {
-	if p.Field.allAttributes {
-		entry.Attributes = nil
-		return nil
-	}
-
-	if p.Field.allResource {
-		entry.Resource = nil
-		return nil
-	}
-
-	_, exist := entry.Delete(p.Field.Field)
-	if !exist {
-		return fmt.Errorf("remove: field does not exist: %s", p.Field.Field.String())
-	}
-	return nil
 }
