+++ pkg/stanza/operator/transformer/remove/transformer_test.go
